wip: move Palette out of zellij-tile
