+++ Cargo.toml
@@ -37,6 +37,7 @@ lazy_static = "1.4.0"
 wasmer = "1.0.0"
 wasmer-wasi = "1.0.0"
 interprocess = "1.0.1"
+xrdb = "0.1.1"
 colors-transform = "0.2.5"
 zellij-tile = { path = "zellij-tile/", version = "0.5.0" }
