+++ zellij-tile/Cargo.toml
@@ -11,5 +11,3 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 strum = "0.20.0"
 strum_macros = "0.20.0"
-xrdb = "0.1.1"
-colors-transform = "0.2.5"
